moved loading indicator to its own component and fixed its position
