+++ src/components/Root.tsx
@@ -5,27 +5,15 @@ import {
   clearCollection,
   fetchSearchDescription
 } from "../actions";
+import LoadingIndicator from "./LoadingIndicator";
 import Collection from "./Collection";
 import UrlForm from "./UrlForm";
 
 export class Root extends React.Component<RootProps, any> {
   render(): JSX.Element {
-    let loadingWidth = 200;
-    let loadingStyle = {
-      position: "absolute",
-      top: "50%",
-      left: "50%",
-      width: `${loadingWidth}px`,
-      marginTop: "-10px",
-      marginLeft: `-${loadingWidth / 2}px`,
-      padding: "30px",
-      backgroundColor: "#bbb",
-      textAlign: "center"
-    };
-
     return (
       <div className="browser">
-        { this.props.isFetching && <h1 className="loading" style={loadingStyle}>LOADING</h1> }
+        { this.props.isFetching && <LoadingIndicator /> }
 
         { this.props.collectionData ?
           <Collection
